Refactor after simple flow removal
removing common file, and make pac file smaller by
creating secrets, pac_repository and pipeline file
which contain corresponding methods

STONEBLD-2612
