+++ controllers/component_build_controller_pipeline.go
@@ -1,0 +1,534 @@
+/*
+Copyright 2021-2024 Red Hat, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package controllers
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+
+	"github.com/google/go-containerregistry/pkg/authn"
+	appstudiov1alpha1 "github.com/konflux-ci/application-api/api/v1alpha1"
+	tektonapi "github.com/tektoncd/pipeline/pkg/apis/pipeline/v1"
+	tektonapi_v1beta1 "github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
+	oci "github.com/tektoncd/pipeline/pkg/remote/oci"
+
+	"sigs.k8s.io/yaml"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/errors"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
+	ctrllog "sigs.k8s.io/controller-runtime/pkg/log"
+
+	"github.com/konflux-ci/build-service/pkg/boerrors"
+	. "github.com/konflux-ci/build-service/pkg/common"
+	gp "github.com/konflux-ci/build-service/pkg/git/gitprovider"
+	l "github.com/konflux-ci/build-service/pkg/logs"
+)
+
+type BuildPipeline struct {
+	Name   string `json:"name,omitempty"`
+	Bundle string `json:"bundle,omitempty"`
+}
+
+type pipelineConfig struct {
+	DefaultPipelineName string          `yaml:"default-pipeline-name"`
+	Pipelines           []BuildPipeline `yaml:"pipelines"`
+}
+
+func (r *ComponentBuildReconciler) ensurePipelineServiceAccount(ctx context.Context, namespace string) (*corev1.ServiceAccount, error) {
+	log := ctrllog.FromContext(ctx)
+
+	pipelinesServiceAccount := &corev1.ServiceAccount{}
+	err := r.Client.Get(ctx, types.NamespacedName{Name: buildPipelineServiceAccountName, Namespace: namespace}, pipelinesServiceAccount)
+	if err != nil {
+		if !errors.IsNotFound(err) {
+			log.Error(err, fmt.Sprintf("Failed to read service account %s in namespace %s", buildPipelineServiceAccountName, namespace), l.Action, l.ActionView)
+			return nil, err
+		}
+		// Create service account for the build pipeline
+		buildPipelineSA := corev1.ServiceAccount{
+			ObjectMeta: metav1.ObjectMeta{
+				Name:      buildPipelineServiceAccountName,
+				Namespace: namespace,
+			},
+		}
+		if err := r.Client.Create(ctx, &buildPipelineSA); err != nil {
+			log.Error(err, fmt.Sprintf("Failed to create service account %s in namespace %s", buildPipelineServiceAccountName, namespace), l.Action, l.ActionAdd)
+			return nil, err
+		}
+		return r.ensurePipelineServiceAccount(ctx, namespace)
+	}
+	return pipelinesServiceAccount, nil
+}
+
+// generatePaCPipelineRunConfigs generates PipelineRun YAML configs for given component.
+// The generated PipelineRun Yaml content are returned in byte string and in the order of push and pull request.
+func (r *ComponentBuildReconciler) generatePaCPipelineRunConfigs(ctx context.Context, component *appstudiov1alpha1.Component, gitClient gp.GitProviderClient, pacTargetBranch string) ([]byte, []byte, error) {
+	log := ctrllog.FromContext(ctx)
+
+	var pipelineName string
+	var pipelineBundle string
+	var pipelineRef *tektonapi.PipelineRef
+	var err error
+
+	// no need to check error because it would fail already in Reconcile
+	pipelineRef, _ = r.GetBuildPipelineFromComponentAnnotation(ctx, component)
+	pipelineName, pipelineBundle, err = getPipelineNameAndBundle(pipelineRef)
+	if err != nil {
+		return nil, nil, err
+	}
+	log.Info(fmt.Sprintf("Selected %s pipeline from %s bundle for %s component",
+		pipelineName, pipelineBundle, component.Name),
+		l.Audit, "true")
+
+	// Get pipeline from the bundle to be expanded to the PipelineRun
+	pipelineSpec, err := retrievePipelineSpec(ctx, pipelineBundle, pipelineName)
+	if err != nil {
+		r.EventRecorder.Event(component, "Warning", "ErrorGettingPipelineFromBundle", err.Error())
+		return nil, nil, err
+	}
+
+	pipelineRunOnPush, err := generatePaCPipelineRunForComponent(component, pipelineSpec, pacTargetBranch, gitClient, false)
+	if err != nil {
+		return nil, nil, err
+	}
+	pipelineRunOnPushYaml, err := yaml.Marshal(pipelineRunOnPush)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	pipelineRunOnPR, err := generatePaCPipelineRunForComponent(component, pipelineSpec, pacTargetBranch, gitClient, true)
+	if err != nil {
+		return nil, nil, err
+	}
+	pipelineRunOnPRYaml, err := yaml.Marshal(pipelineRunOnPR)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return pipelineRunOnPushYaml, pipelineRunOnPRYaml, nil
+}
+
+// retrievePipelineSpec retrieves pipeline definition with given name from the given bundle.
+func retrievePipelineSpec(ctx context.Context, bundleUri, pipelineName string) (*tektonapi.PipelineSpec, error) {
+	log := ctrllog.FromContext(ctx)
+
+	var obj runtime.Object
+	var err error
+	resolver := oci.NewResolver(bundleUri, authn.DefaultKeychain)
+
+	if obj, _, err = resolver.Get(ctx, "pipeline", pipelineName); err != nil {
+		return nil, err
+	}
+
+	var pipelineSpec tektonapi.PipelineSpec
+
+	if v1beta1Pipeline, ok := obj.(tektonapi_v1beta1.PipelineObject); ok {
+		v1beta1PipelineSpec := v1beta1Pipeline.PipelineSpec()
+		log.Info("Converting from v1beta1 to v1", "PipelineName", pipelineName, "Bundle", bundleUri)
+		err := v1beta1PipelineSpec.ConvertTo(ctx, &pipelineSpec, &metav1.ObjectMeta{})
+		if err != nil {
+			return nil, boerrors.NewBuildOpError(
+				boerrors.EPipelineConversionFailed,
+				fmt.Errorf("pipeline %s from bundle %s: failed to convert from v1beta1 to v1: %w", pipelineName, bundleUri, err),
+			)
+		}
+	} else if v1Pipeline, ok := obj.(*tektonapi.Pipeline); ok {
+		pipelineSpec = v1Pipeline.PipelineSpec()
+	} else {
+		return nil, boerrors.NewBuildOpError(
+			boerrors.EPipelineRetrievalFailed,
+			fmt.Errorf("failed to extract pipeline %s from bundle %s", pipelineName, bundleUri),
+		)
+	}
+
+	return &pipelineSpec, nil
+}
+
+// GetBuildPipelineFromComponentAnnotation parses pipeline annotation on component and returns build pipeline
+func (r *ComponentBuildReconciler) GetBuildPipelineFromComponentAnnotation(ctx context.Context, component *appstudiov1alpha1.Component) (*tektonapi.PipelineRef, error) {
+	buildPipeline, err := readBuildPipelineAnnotation(component)
+	if err != nil {
+		return nil, err
+	}
+	if buildPipeline == nil {
+		err := fmt.Errorf("missing or empty pipeline annotation: %s, will add default one to the component", component.Annotations[defaultBuildPipelineAnnotation])
+		return nil, boerrors.NewBuildOpError(boerrors.EMissingPipelineAnnotation, err)
+	}
+	if buildPipeline.Bundle == "" || buildPipeline.Name == "" {
+		err = fmt.Errorf("missing name or bundle in pipeline annotation: name=%s bundle=%s", buildPipeline.Name, buildPipeline.Bundle)
+		return nil, boerrors.NewBuildOpError(boerrors.EWrongPipelineAnnotation, err)
+	}
+	finalBundle := buildPipeline.Bundle
+
+	if buildPipeline.Bundle == "latest" {
+		pipelinesConfigMap := &corev1.ConfigMap{}
+		if err := r.Client.Get(ctx, types.NamespacedName{Name: buildPipelineConfigMapResourceName, Namespace: BuildServiceNamespaceName}, pipelinesConfigMap); err != nil {
+			if errors.IsNotFound(err) {
+				return nil, boerrors.NewBuildOpError(boerrors.EBuildPipelineConfigNotDefined, err)
+			}
+			return nil, err
+		}
+
+		buildPipelineData := &pipelineConfig{}
+		if err := yaml.Unmarshal([]byte(pipelinesConfigMap.Data[buildPipelineConfigName]), buildPipelineData); err != nil {
+			return nil, boerrors.NewBuildOpError(boerrors.EBuildPipelineConfigNotValid, err)
+		}
+
+		for _, pipeline := range buildPipelineData.Pipelines {
+			if pipeline.Name == buildPipeline.Name {
+				finalBundle = pipeline.Bundle
+				break
+			}
+		}
+
+		// requested pipeline was not found in configMap
+		if finalBundle == "latest" {
+			err = fmt.Errorf("invalid pipeline name in pipeline annotation: name=%s", buildPipeline.Name)
+			return nil, boerrors.NewBuildOpError(boerrors.EBuildPipelineInvalid, err)
+		}
+	}
+
+	pipelineRef := &tektonapi.PipelineRef{
+		ResolverRef: tektonapi.ResolverRef{
+			Resolver: "bundles",
+			Params: []tektonapi.Param{
+				{Name: "name", Value: *tektonapi.NewStructuredValues(buildPipeline.Name)},
+				{Name: "bundle", Value: *tektonapi.NewStructuredValues(finalBundle)},
+				{Name: "kind", Value: *tektonapi.NewStructuredValues("pipeline")},
+			},
+		},
+	}
+	return pipelineRef, nil
+}
+
+func readBuildPipelineAnnotation(component *appstudiov1alpha1.Component) (*BuildPipeline, error) {
+	if component.Annotations == nil {
+		return nil, nil
+	}
+
+	requestedPipeline, requestedPipelineExists := component.Annotations[defaultBuildPipelineAnnotation]
+	if requestedPipelineExists && requestedPipeline != "" {
+		buildPipeline := &BuildPipeline{}
+		buildPipelineBytes := []byte(requestedPipeline)
+
+		if err := json.Unmarshal(buildPipelineBytes, buildPipeline); err != nil {
+			return nil, boerrors.NewBuildOpError(boerrors.EFailedToParsePipelineAnnotation, err)
+		}
+		return buildPipeline, nil
+	}
+	return nil, nil
+}
+
+// SetDefaultBuildPipelineComponentAnnotation sets default build pipeline to component pipeline annotation
+func (r *ComponentBuildReconciler) SetDefaultBuildPipelineComponentAnnotation(ctx context.Context, component *appstudiov1alpha1.Component) error {
+	log := ctrllog.FromContext(ctx)
+	pipelinesConfigMap := &corev1.ConfigMap{}
+
+	if err := r.Client.Get(ctx, types.NamespacedName{Name: buildPipelineConfigMapResourceName, Namespace: BuildServiceNamespaceName}, pipelinesConfigMap); err != nil {
+		if errors.IsNotFound(err) {
+			return boerrors.NewBuildOpError(boerrors.EBuildPipelineConfigNotDefined, err)
+		}
+		return err
+	}
+
+	buildPipelineData := &pipelineConfig{}
+	if err := yaml.Unmarshal([]byte(pipelinesConfigMap.Data[buildPipelineConfigName]), buildPipelineData); err != nil {
+		return boerrors.NewBuildOpError(boerrors.EBuildPipelineConfigNotValid, err)
+	}
+
+	pipelineAnnotation := fmt.Sprintf("{\"name\":\"%s\",\"bundle\":\"%s\"}", buildPipelineData.DefaultPipelineName, "latest")
+	if component.Annotations == nil {
+		component.Annotations = make(map[string]string)
+	}
+	component.Annotations[defaultBuildPipelineAnnotation] = pipelineAnnotation
+
+	if err := r.Client.Update(ctx, component); err != nil {
+		log.Error(err, fmt.Sprintf("failed to update component with default pipeline annotation %s", defaultBuildPipelineAnnotation))
+		return err
+	}
+	log.Info(fmt.Sprintf("updated component with default pipeline annotation %s", defaultBuildPipelineAnnotation))
+	return nil
+}
+
+// generatePaCPipelineRunForComponent returns pipeline run definition to build component source with.
+// Generated pipeline run contains placeholders that are expanded by Pipeline-as-Code.
+func generatePaCPipelineRunForComponent(
+	component *appstudiov1alpha1.Component,
+	pipelineSpec *tektonapi.PipelineSpec,
+	pacTargetBranch string,
+	gitClient gp.GitProviderClient,
+	onPull bool) (*tektonapi.PipelineRun, error) {
+
+	if pacTargetBranch == "" {
+		return nil, fmt.Errorf("target branch can't be empty for generating PaC PipelineRun for: %v", component)
+	}
+	pipelineCelExpression, err := generateCelExpressionForPipeline(component, gitClient, pacTargetBranch, onPull)
+	if err != nil {
+		return nil, fmt.Errorf("failed to generate cel expression for pipeline: %w", err)
+	}
+	repoUrl := component.Spec.Source.GitSource.URL
+
+	annotations := map[string]string{
+		"pipelinesascode.tekton.dev/max-keep-runs": "3",
+		"build.appstudio.redhat.com/target_branch": "{{target_branch}}",
+		pacCelExpressionAnnotationName:             pipelineCelExpression,
+		gitCommitShaAnnotationName:                 "{{revision}}",
+		gitRepoAtShaAnnotationName:                 gitClient.GetBrowseRepositoryAtShaLink(repoUrl, "{{revision}}"),
+	}
+	labels := map[string]string{
+		ApplicationNameLabelName:                component.Spec.Application,
+		ComponentNameLabelName:                  component.Name,
+		"pipelines.appstudio.openshift.io/type": "build",
+	}
+
+	imageRepo := getContainerImageRepositoryForComponent(component)
+
+	var pipelineName string
+	var proposedImage string
+	if onPull {
+		annotations["build.appstudio.redhat.com/pull_request_number"] = "{{pull_request_number}}"
+		pipelineName = component.Name + pipelineRunOnPRSuffix
+		proposedImage = imageRepo + ":on-pr-{{revision}}"
+	} else {
+		pipelineName = component.Name + pipelineRunOnPushSuffix
+		proposedImage = imageRepo + ":{{revision}}"
+	}
+
+	params := []tektonapi.Param{
+		{Name: "git-url", Value: tektonapi.ParamValue{Type: "string", StringVal: "{{source_url}}"}},
+		{Name: "revision", Value: tektonapi.ParamValue{Type: "string", StringVal: "{{revision}}"}},
+		{Name: "output-image", Value: tektonapi.ParamValue{Type: "string", StringVal: proposedImage}},
+	}
+	if onPull {
+		prImageExpiration := os.Getenv(PipelineRunOnPRExpirationEnvVar)
+		if prImageExpiration == "" {
+			prImageExpiration = PipelineRunOnPRExpirationDefault
+		}
+		params = append(params, tektonapi.Param{Name: "image-expires-after", Value: tektonapi.ParamValue{Type: "string", StringVal: prImageExpiration}})
+	}
+
+	if component.Spec.Source.GitSource.DockerfileURL != "" {
+		params = append(params, tektonapi.Param{Name: "dockerfile", Value: tektonapi.ParamValue{Type: "string", StringVal: component.Spec.Source.GitSource.DockerfileURL}})
+	} else {
+		params = append(params, tektonapi.Param{Name: "dockerfile", Value: tektonapi.ParamValue{Type: "string", StringVal: "Dockerfile"}})
+	}
+	pathContext := getPathContext(component.Spec.Source.GitSource.Context, "")
+	if pathContext != "" {
+		params = append(params, tektonapi.Param{Name: "path-context", Value: tektonapi.ParamValue{Type: "string", StringVal: pathContext}})
+	}
+
+	pipelineRunWorkspaces := createWorkspaceBinding(pipelineSpec.Workspaces)
+
+	pipelineRun := &tektonapi.PipelineRun{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "PipelineRun",
+			APIVersion: "tekton.dev/v1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Name:        pipelineName,
+			Namespace:   component.Namespace,
+			Labels:      labels,
+			Annotations: annotations,
+		},
+		Spec: tektonapi.PipelineRunSpec{
+			PipelineSpec: pipelineSpec,
+			Params:       params,
+			Workspaces:   pipelineRunWorkspaces,
+		},
+	}
+
+	return pipelineRun, nil
+}
+
+func getPathContext(gitContext, dockerfileContext string) string {
+	if gitContext == "" && dockerfileContext == "" {
+		return ""
+	}
+	separator := string(filepath.Separator)
+	path := filepath.Join(gitContext, dockerfileContext)
+	path = filepath.Clean(path)
+	path = strings.TrimPrefix(path, separator)
+	return path
+}
+
+func createWorkspaceBinding(pipelineWorkspaces []tektonapi.PipelineWorkspaceDeclaration) []tektonapi.WorkspaceBinding {
+	pipelineRunWorkspaces := []tektonapi.WorkspaceBinding{}
+	for _, workspace := range pipelineWorkspaces {
+		switch workspace.Name {
+		case "workspace":
+			pipelineRunWorkspaces = append(pipelineRunWorkspaces,
+				tektonapi.WorkspaceBinding{
+					Name:                workspace.Name,
+					VolumeClaimTemplate: generateVolumeClaimTemplate(),
+				})
+		case "git-auth":
+			pipelineRunWorkspaces = append(pipelineRunWorkspaces,
+				tektonapi.WorkspaceBinding{
+					Name:   workspace.Name,
+					Secret: &corev1.SecretVolumeSource{SecretName: "{{ git_auth_secret }}"},
+				})
+		}
+	}
+	return pipelineRunWorkspaces
+}
+
+func generateVolumeClaimTemplate() *corev1.PersistentVolumeClaim {
+	return &corev1.PersistentVolumeClaim{
+		Spec: corev1.PersistentVolumeClaimSpec{
+			AccessModes: []corev1.PersistentVolumeAccessMode{
+				"ReadWriteOnce",
+			},
+			Resources: corev1.VolumeResourceRequirements{
+				Requests: corev1.ResourceList{
+					"storage": resource.MustParse("1Gi"),
+				},
+			},
+		},
+	}
+}
+
+// generateCelExpressionForPipeline generates value for pipelinesascode.tekton.dev/on-cel-expression annotation
+// in order to have better flexibility with git events filtering.
+// Examples of returned values:
+// event == "push" && target_branch == "main"
+// event == "pull_request" && target_branch == "my-branch" && ( "component-src-dir/***".pathChanged() || ".tekton/pipeline.yaml".pathChanged() || "dockerfiles/my-component/Dockerfile".pathChanged() )
+func generateCelExpressionForPipeline(component *appstudiov1alpha1.Component, gitClient gp.GitProviderClient, targetBranch string, onPull bool) (string, error) {
+	eventType := "push"
+	if onPull {
+		eventType = "pull_request"
+	}
+	eventCondition := fmt.Sprintf(`event == "%s"`, eventType)
+
+	targetBranchCondition := fmt.Sprintf(`target_branch == "%s"`, targetBranch)
+
+	// Set path changed event filtering only for Components that are stored within a directory of the git repository.
+	// Also, we have to rebuild everything on push events, so applying the filter only to pull request pipeline.
+	pathChangedSuffix := ""
+	if onPull && component.Spec.Source.GitSource.Context != "" && component.Spec.Source.GitSource.Context != "/" && component.Spec.Source.GitSource.Context != "./" && component.Spec.Source.GitSource.Context != "." {
+		contextDir := component.Spec.Source.GitSource.Context
+		if !strings.HasSuffix(contextDir, "/") {
+			contextDir += "/"
+		}
+
+		// If a Dockerfile is defined for the Component,
+		// we should rebuild the Component if the Dockerfile has been changed.
+		dockerfilePathChangedSuffix := ""
+		dockerfile := component.Spec.Source.GitSource.DockerfileURL
+		if dockerfile != "" {
+			// Ignore dockerfile that is not stored in the same git repository but downloaded by an URL.
+			if !strings.Contains(dockerfile, "://") {
+				// dockerfile could be relative to the context directory or repository root.
+				// To avoid unessesary builds, it's required to pass absolute path to the Dockerfile.
+				repoUrl := component.Spec.Source.GitSource.URL
+				branch := component.Spec.Source.GitSource.Revision
+				dockerfilePath := contextDir + dockerfile
+				isDockerfileInContextDir, err := gitClient.IsFileExist(repoUrl, branch, dockerfilePath)
+				if err != nil {
+					return "", err
+				}
+				// If the Dockerfile is inside context directory, no changes to event filter needed.
+				if !isDockerfileInContextDir {
+					// Pipelines as Code doesn't match path if it starts from /
+					dockerfileAbsolutePath := strings.TrimPrefix(dockerfile, "/")
+					dockerfilePathChangedSuffix = fmt.Sprintf(`|| "%s".pathChanged() `, dockerfileAbsolutePath)
+				}
+			}
+		}
+
+		pullPipelineFileName := component.Name + "-" + pipelineRunOnPRFilename
+		pathChangedSuffix = fmt.Sprintf(` && ( "%s***".pathChanged() || ".tekton/%s".pathChanged() %s)`, contextDir, pullPipelineFileName, dockerfilePathChangedSuffix)
+	}
+
+	return fmt.Sprintf("%s && %s%s", eventCondition, targetBranchCondition, pathChangedSuffix), nil
+}
+
+func getContainerImageRepositoryForComponent(component *appstudiov1alpha1.Component) string {
+	if component.Spec.ContainerImage != "" {
+		return getContainerImageRepository(component.Spec.ContainerImage)
+	}
+	imageRepo, _, err := getComponentImageRepoAndSecretNameFromImageAnnotation(component)
+	if err == nil && imageRepo != "" {
+		return imageRepo
+	}
+	return ""
+}
+
+// getContainerImageRepository removes tag or SHA has from container image reference
+func getContainerImageRepository(image string) string {
+	if strings.Contains(image, "@") {
+		// registry.io/user/image@sha256:586ab...d59a
+		return strings.Split(image, "@")[0]
+	}
+	// registry.io/user/image:tag
+	return strings.Split(image, ":")[0]
+}
+
+// getComponentImageRepoAndSecretNameFromImageAnnotation parses image.redhat.com/image annotation
+// for image repository and secret name to access it.
+// If image.redhat.com/image is not set, the procedure returns empty values.
+func getComponentImageRepoAndSecretNameFromImageAnnotation(component *appstudiov1alpha1.Component) (string, string, error) {
+	type RepositoryInfo struct {
+		Image  string `json:"image"`
+		Secret string `json:"secret"`
+	}
+
+	var repoInfo RepositoryInfo
+	if imageRepoDataJson, exists := component.Annotations[ImageRepoAnnotationName]; exists {
+		if err := json.Unmarshal([]byte(imageRepoDataJson), &repoInfo); err != nil {
+			return "", "", boerrors.NewBuildOpError(boerrors.EFailedToParseImageAnnotation, err)
+		}
+		return repoInfo.Image, repoInfo.Secret, nil
+	}
+	return "", "", nil
+}
+
+func getPipelineNameAndBundle(pipelineRef *tektonapi.PipelineRef) (string, string, error) {
+	if pipelineRef.Resolver != "" && pipelineRef.Resolver != "bundles" {
+		return "", "", boerrors.NewBuildOpError(
+			boerrors.EUnsupportedPipelineRef,
+			fmt.Errorf("unsupported Tekton resolver %q", pipelineRef.Resolver),
+		)
+	}
+
+	name := pipelineRef.Name
+	var bundle string
+
+	for _, param := range pipelineRef.Params {
+		switch param.Name {
+		case "name":
+			name = param.Value.StringVal
+		case "bundle":
+			bundle = param.Value.StringVal
+		}
+	}
+
+	if name == "" || bundle == "" {
+		return "", "", boerrors.NewBuildOpError(
+			boerrors.EMissingParamsForBundleResolver,
+			fmt.Errorf("missing name or bundle in pipelineRef: name=%s bundle=%s", name, bundle),
+		)
+	}
+
+	return name, bundle, nil
+}
